+++ src/commands.rs
@@ -1,7 +1,8 @@
 #[macro_use]
 pub(crate) mod macros;
 
-mod from_structured_data;
+mod from_delimited_data;
+mod to_delimited_data;
 
 pub(crate) mod append;
 pub(crate) mod args;
